Refactor WatcherService into a property of ProjectExplorerViewModel to ensure consistent source of truth and state flow.
