+++ WolvenKit.App/Services/AppScriptService.cs
@@ -68,7 +68,7 @@ public partial class AppScriptService : ScriptService
     {
         GetProjectExplorerViewModel()?.SuspendFileWatcher();
         await ExecuteAsync(code, DefaultHostObject, null, enableDebugging);
-        GetProjectExplorerViewModel()?.ResumeFileWatcher_AndReloadProject();
+        GetProjectExplorerViewModel()?.ResumeWatcher_AndReloadProject();
     }
 
     public void SetAppViewModel(AppViewModel appViewModel) => _wkit.AppViewModel = appViewModel;
